+++ Code/Client/users/src/pages/car-rental/carFeatureModel/carFeature.css
@@ -1,13 +1,45 @@
-.modalStyle{
-    height: 100%;
+.modalStyle {
+    height: 60%;
     width: 100%;
     background-color: rgb(180, 180, 180);
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    background-color: #fff;
+    padding: 30px;
+    border-radius: 8px;
+    box-shadow: 0 4px 8px #6c757d;
+    max-width: 500px; /* Adjust as needed */
+    width: 100%;
+    z-index: 1000;
+    overflow-y: hidden;
 }
 
-.heading{
-    font-size: 25px;
-    font-weight: 700;
+.overlayStyle {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.3);
+    z-index: 999;
+    backdrop-filter: blur(2px);
+}
+
+.closeButtonStyle {
+    cursor: pointer;
+    position: absolute;
+    top: 27px;
+    right: 30px;
+    font-size: 2rem;
+}
+
+.modalStyle h4{
+    font-size: 24px;
+    font-weight: 600;
     color: blueviolet;
+    padding-bottom: 10px;
 }
 
 .accordion-section {
@@ -18,6 +50,7 @@
     display: flex;
     justify-content: space-between;
     align-items: center;
+    font-size: 15px;
     padding: 10px;
     background-color: #f1f1f1;
     cursor: pointer;
@@ -48,4 +81,10 @@
 
 .accordion-section.active .accordion-details {
     max-height: 400px;
+}
+
+.feature_btn{
+    margin-top: 30px;
+    display: flex;
+    justify-content: flex-end;
 }
